added generic config
